+++ HISTORY.rst
@@ -10,6 +10,7 @@ History
     * Expose initial password on UserProfile creation (ericof).
     * Set timezone on new and updated OrderLocations (ericof).
     * Fix remove_availability transition: now the new assignment is created after cancel the old one (rudaporto).
+    * Scout dashboard now support links on projects (rudaporto).
 
 
 2.0.3 (2017-02-14)